requirements.txt
beautifulsoup4==4.10.0
jieba==0.39
numpy==1.18.5
pandas==0.25.3
Pillow==9.0.0
python_dateutil==2.8.2
requests==2.25.1
requests_html==0.6.6
SQLAlchemy==1.3.24
SQLObject==3.9.1
torch==1.0.0
torchtext==0.3.1
tqdm==4.62.3
wordcloud==1.6.0
